requirements.txt
streamlit
anthropic
PyPDF2
sentence-transformers
faiss-cpu
numpy
torch
transformers
